added admin refactor
