+++ web/src/pages/admin/AdminRoute.js
@@ -2,8 +2,8 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { Route, Redirect } from 'react-router-dom';
 import { connect } from 'react-redux';
-import { getIsFedAdmin } from '../reducers/user.selector';
-import NoMatch from '../components/NoMatch';
+import { getIsFedAdmin } from '../../reducers/user.selector';
+import NoMatch from '../../components/NoMatch';
 
 const AdminRoute = ({
   authenticated,
