+++ web/src/pages/admin/ConfirmationDialog.js
